Pass restricted domain as string when only one
restrictCreationByEmailDomain accepts either a function or a string.
When it is a string, it will be used by some of the OAuth integrations
(like google) to proactively restrict logins.

For my use case this means that employees are automatically asked by
google to login with their work credentials instead of first having to
select between their personal and work accounts.