[FIX] Improving consistency of UX (#8796)
* Capitalized letters of 'Reset Section Settings'

Improves the consistency of UX; every other field title has capitalized letters, and this stands out and looks a little unpolished.

* [FIX] Capitalized more words in Administration

More consistent UX; far easier to read.

* [FIX] Capitalized more words in Administration

Looking even better. :- )

* [FIX] Capitalized field titles in Message

Capitalized field titles in Message in Administration.

* [FIX] Capitalized field titles in SlackBridge

Improved consistency of titles in SlackBridge in Administration.

* [FIX] Capitalized 1st letter of SAML field titles

Improved consistency of field titles of SAML in Administration.

* [FIX] Capitalized 1st letter of Message titles

Improved consistency of field titles in Message in Administration.

* [FIX] Capitalized 1st letter of Push titles

Improved consistency by capitalizing the first letter of the field titles in Push in Administration.

* [FIX] Capitalized 1st letter of titles in Livechat

Improved consistency of field titles in Livechat in Administration by capitalizing the first letters of each word.

* [FIX] Capitalized 1st letter of titles on Livechat

Further improved UX consistency of expandable configuration options in Livechat in Administration by capitalizing the first letter of each word in the field titles.

* [FIX] Capitalize 1st letter of titles on LDAP

Capitalized first letter of each word in the field titles on LDAP in Administration to improve consistency of UX.

* [FIX] Changed 'Videocall' to 'Video Call'

Switching between the two wasn't great for UX consistency. We'll have 'Direct Messages' and 'Video Calls'. :- )

* [FIX] Capitalized 1st letter of titles on Layout

Capitalized the first letter of each word in the field titles of Layout in Administration. Also corrected grammar of something. :- )

* [FIX] Capitalized 1st letter of titles on Internal Hubot

Capitalized first letter of each field title on Internal Hubot on Administration for the purposes of consistency of UX.

* [FIX] Capitalized 1st letter of titles on General

Capitalized the first letter of each word in the field titles of General in Administration for the purpose of consistency in terms of UX.

* [FIX] Capitalized 1st letter of File Upload titles

Capitalized the first letter of each word in the field titles of File Upload in Administration for the purpose of making the UX more consistent.

* [FIX] Capitalized 1st letter of titles in Email

Capitalized first letter of each word of the field titles in Email in Administration for the sake of UX consistency.

* [FIX] Capitalized 1st letters of titles in CAS

Capitalized first letters of field titles in CAS in Administration for the sake of UX consistency.

* [FIX] Capitalized 1st letter of Analytics titles

Capitalized first letter of each word of the field titles in Analytics in Administration for the sake of UX consistency.

* [FIX] Improved consistency of Accounts UX

* [FIX] Changed a mention of 'conversations'

Changed it to 'Direct Messages'.

* Changed a mention of 'conversation'

Changed another mention of 'Conversation' to 'Direct Message'.

* [FIX] Changed 'Show all' to 'Show All'

Looks far better in terms of UX.

* [FIX] Improved consistency of UX

Capitalized letters on field titles in Preferences.

* [FIX] Improved consistency of actions

Was the only thing not capitalized in this menu and stuck out like a sore thumb. :- )

* [FIX] Changed 'Hide room' to 'Hide Room'

Didn't match the rest of the action menus.

* [FIX] Added a word to modal

The 'back' action of 'View All' wasn't exactly intuitive as to what you were viewing; made it seem as if you were viewing more methods of interaction with that user rather than returning to a list of members, so changed it to 'View All Members'.

* [FIX] Grammar

:- )

* [FIX] Capitalized 'Name' in 'Channel Name'

Only field title that wasn't capitalized on the 'Create a New Channel' area.

* [FIX] Dropped unnecessary periods to clean up UX

Periods were unnecessary, and looks far cleaner without. Using periods isn't commonly done for this part of the UX.