[FIX] Resolve 'app already exists' error on app update (#17544)
* fix: resolve 'app already exists' error on app update

The 'Apps' and 'Marketplace' admin pages were using the function
`installApp` from the `AppClientOrchestrator` class instead of `updateApp` to
update apps. As a result, the 'app already exists' error was being thrown and
apps were not being updated.

* fix: remove `install app` handle from apps admin page

As the `install app` handle is only needed in the marketplace page,
there's no need to have it in the apps admin page.